Add implementation to service method for creating a track

+++ backend/song-service/src/main/java/edu/personal/grailsfm/songservice/entity/Track.java
@@ -37,7 +37,7 @@
         private String description;
 
         @Field(name = "duration")
-        private Integer duration;
+        private Float duration;
 
         @Field(name = "genres")
         private List<String> genres;
